someone said to use Rc, i think thats wrong #16
